requirements.txt
Django==5.1.3
Jinja2==3.1.4
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.18.0
Willow==1.9.0
anyascii==0.3.2
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
defusedxml==0.7.1
django-allauth==65.2.0
django-browser-reload==1.17.0
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==6.1.0
django-tailwind==3.8.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
openpyxl==3.1.5
pillow==11.0.0
pillow_heif==0.20.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
rich==13.9.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
telepath==0.3.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
wagtail==6.3.1
